feat: Refactor UserRepository and UserService
The UserRepository interface has been updated to include a new method findDistinctAccessByUserId, which retrieves a distinct list of Access entities based on the provided user ID. Additionally, the UserService class has been modified to use the new method when updating the user's access. These changes improve the functionality and efficiency of the code.